change indices in AST to usize, since they point into Vecs (so we no longer need to cast between u32 <-> usize)

+++ instrumentation/own/wasm-instrumentation/src/ast.rs
@@ -183,25 +183,25 @@ pub enum ExportType {
 /* Indices */
 
 #[derive(WasmBinary, Debug, PartialEq, PartialOrd)]
-pub struct TypeIdx(pub Leb128<u32>);
+pub struct TypeIdx(pub Leb128<usize>);
 
 #[derive(WasmBinary, Debug, PartialEq, PartialOrd)]
-pub struct FunctionIdx(pub Leb128<u32>);
+pub struct FunctionIdx(pub Leb128<usize>);
 
 #[derive(WasmBinary, Debug, PartialEq, PartialOrd)]
-pub struct TableIdx(pub Leb128<u32>);
+pub struct TableIdx(pub Leb128<usize>);
 
 #[derive(WasmBinary, Debug, PartialEq, PartialOrd)]
-pub struct MemoryIdx(pub Leb128<u32>);
+pub struct MemoryIdx(pub Leb128<usize>);
 
 #[derive(WasmBinary, Debug, PartialEq, PartialOrd)]
-pub struct GlobalIdx(pub Leb128<u32>);
+pub struct GlobalIdx(pub Leb128<usize>);
 
 #[derive(WasmBinary, Debug, PartialEq, PartialOrd)]
-pub struct LocalIdx(pub Leb128<u32>);
+pub struct LocalIdx(pub Leb128<usize>);
 
 #[derive(WasmBinary, Debug, PartialEq, PartialOrd)]
-pub struct LabelIdx(pub Leb128<u32>);
+pub struct LabelIdx(pub Leb128<usize>);
 
 
 /* Code */
